moved more methods, renamed AITools to AIUtils
