+++ game/src/contrib/utils/components/ai/AIUtils.java
@@ -7,15 +7,10 @@ import core.Game;
 import core.components.PositionComponent;
 import core.components.VelocityComponent;
 import core.level.Tile;
-import core.level.utils.Coordinate;
 import core.level.utils.LevelUtils;
-import core.utils.Point;
 import core.utils.components.MissingComponentException;
 
-import java.util.*;
-
-public class AITools {
-    private static final Random random = new Random();
+public class AIUtils {
 
     /**
      * Sets the velocity of the passed entity so that it takes the next necessary step to get to the
@@ -68,107 +63,6 @@ public class AITools {
                 case E -> vc.currentXVelocity(vc.xVelocity());
                 case W -> vc.currentXVelocity(-vc.xVelocity());
             }
-    }
-
-    /**
-     * Get all tiles within a specified range around a given center point.
-     *
-     * <p>The range is determined by the provided radius.
-     *
-     * <p>The tile at the given point will be part of the list as well.
-     *
-     * @param center The center point around which the tiles are considered.
-     * @param radius The radius within which the tiles should be located.
-     * @return List of tiles in the given radius around the center point.
-     */
-    public static List<Tile> tilesInRange(final Point center, final float radius) {
-        List<Tile> tiles = new ArrayList<>();
-        for (float x = center.x - radius; x <= center.x + radius; x++) {
-            for (float y = center.y - radius; y <= center.y + radius; y++) {
-                tiles.add(Game.tileAT(new Point(x, y)));
-            }
-        }
-        tiles.removeIf(Objects::isNull);
-        return tiles;
-    }
-
-    /**
-     * Get all accessible tiles within a specified range around a given center point.
-     *
-     * <p>The range is determined by the provided radius.
-     *
-     * <p>The tile at the given point will be part of the list as well, if it is accessible.
-     *
-     * @param center The center point around which the tiles are considered.
-     * @param radius The radius within which the accessible tiles should be located.
-     * @return List of accessible tiles in the given radius around the center point.
-     */
-    public static List<Tile> accessibleTilesInRange(final Point center, final float radius) {
-        List<Tile> tiles = tilesInRange(center, radius);
-        tiles.removeIf(tile -> !tile.isAccessible());
-        return tiles;
-    }
-
-    /**
-     * Get a random accessible tile coordinate within a specified range around a given center point.
-     *
-     * <p>The range is determined by the provided radius.
-     *
-     * <p>The tile at the given point can be the return value as well, if it is accessible.
-     *
-     * @param center The center point around which the tiles are considered.
-     * @param radius The radius within which the accessible tiles should be located.
-     * @return An Optional containing a random Coordinate object representing an accessible tile
-     *     within the range, or an empty Optional if no accessible tiles were found.
-     */
-    public static Optional<Coordinate> randomAccessibleTileCoordinateInRange(
-            final Point center, final float radius) {
-        List<Tile> tiles = accessibleTilesInRange(center, radius);
-        if (tiles.isEmpty()) return Optional.empty();
-        Coordinate newPosition = tiles.get(random.nextInt(tiles.size())).coordinate();
-        return Optional.of(newPosition);
-    }
-
-    /**
-     * Check if two entities are positioned in a specified range from each other.
-     *
-     * @param entity1 The first entity which is considered.
-     * @param entity2 The second entity which is to be searched for in the given range.
-     * @param range The range in which the two entities are positioned from each other.
-     * @return True if the position of the two entities is within the given range, else false
-     */
-    public static boolean entityInRange(
-            final Entity entity1, final Entity entity2, final float range) {
-
-        Point entity1Position =
-                entity1.fetch(PositionComponent.class)
-                        .orElseThrow(
-                                () ->
-                                        MissingComponentException.build(
-                                                entity1, PositionComponent.class))
-                        .position();
-        Point entity2Position =
-                entity2.fetch(PositionComponent.class)
-                        .orElseThrow(
-                                () ->
-                                        MissingComponentException.build(
-                                                entity2, PositionComponent.class))
-                        .position();
-        return Point.inRange(entity1Position, entity2Position, range);
-    }
-
-    /**
-     * Check if the player is in the given range of an entity.
-     *
-     * @param entity Entity whose position specifies the center point.
-     * @param range The range within which the player should be located.
-     * @return True if the position of the player is within the given radius of the position of the
-     *     given entity. If there is no hero, return false.
-     */
-    public static boolean playerInRange(final Entity entity, final float range) {
-
-        Optional<Entity> hero = Game.hero();
-        return hero.filter(value -> entityInRange(entity, value, range)).isPresent();
     }
 
     /**
