+++ game/src/contrib/utils/components/ai/idle/PatrouilleWalk.java
@@ -2,7 +2,7 @@ package contrib.utils.components.ai.idle;
 
 import com.badlogic.gdx.ai.pfa.GraphPath;
 
-import contrib.utils.components.ai.AITools;
+import contrib.utils.components.ai.AIUtils;
 
 import core.Entity;
 import core.Game;
@@ -76,7 +76,7 @@ public class PatrouilleWalk implements Consumer<Entity> {
             return;
         }
 
-        List<Tile> accessibleTiles = AITools.accessibleTilesInRange(center, radius);
+        List<Tile> accessibleTiles = LevelUtils.accessibleTilesInRange(center, radius);
 
         if (accessibleTiles.isEmpty()) {
             return;
@@ -105,18 +105,18 @@ public class PatrouilleWalk implements Consumer<Entity> {
                                         MissingComponentException.build(
                                                 entity, PositionComponent.class));
 
-        if (currentPath != null && !AITools.pathFinished(entity, currentPath)) {
-            if (AITools.pathLeft(entity, currentPath)) {
+        if (currentPath != null && !AIUtils.pathFinished(entity, currentPath)) {
+            if (AIUtils.pathLeft(entity, currentPath)) {
                 currentPath =
                         LevelUtils.calculatePath(
                                 position.position(),
                                 this.checkpoints.get(currentCheckpoint).position());
             }
-            AITools.move(entity, currentPath);
+            AIUtils.move(entity, currentPath);
             return;
         }
 
-        if (currentPath != null && AITools.pathFinished(entity, currentPath)) {
+        if (currentPath != null && AIUtils.pathFinished(entity, currentPath)) {
             frameCounter = 0;
             currentPath = null;
             return;
