+++ game/src/contrib/utils/components/ai/idle/StaticRadiusWalk.java
@@ -2,7 +2,7 @@ package contrib.utils.components.ai.idle;
 
 import com.badlogic.gdx.ai.pfa.GraphPath;
 
-import contrib.utils.components.ai.AITools;
+import contrib.utils.components.ai.AIUtils;
 
 import core.Entity;
 import core.Game;
@@ -38,7 +38,7 @@ public class StaticRadiusWalk implements Consumer<Entity> {
 
     @Override
     public void accept(final Entity entity) {
-        if (path == null || AITools.pathFinishedOrLeft(entity, path)) {
+        if (path == null || AIUtils.pathFinishedOrLeft(entity, path)) {
             if (center == null) {
                 PositionComponent pc =
                         entity.fetch(PositionComponent.class)
@@ -59,7 +59,7 @@ public class StaticRadiusWalk implements Consumer<Entity> {
                                                         entity, PositionComponent.class));
                 currentPosition = pc2.position();
                 newEndTile =
-                        AITools.randomAccessibleTileCoordinateInRange(center, radius)
+                        LevelUtils.randomAccessibleTileCoordinateInRange(center, radius)
                                 .map(Coordinate::toPoint)
                                 // center is the start position of the entity, so it must be
                                 // accessible
@@ -69,6 +69,6 @@ public class StaticRadiusWalk implements Consumer<Entity> {
             }
             currentBreak++;
 
-        } else AITools.move(entity, path);
+        } else AIUtils.move(entity, path);
     }
 }
