+++ game/src/contrib/utils/components/ai/transition/RangeTransition.java
@@ -1,8 +1,7 @@
 package contrib.utils.components.ai.transition;
 
-import contrib.utils.components.ai.AITools;
-
 import core.Entity;
+import core.level.utils.LevelUtils;
 
 import java.util.function.Function;
 
@@ -21,6 +20,6 @@ public class RangeTransition implements Function<Entity, Boolean> {
 
     @Override
     public Boolean apply(final Entity entity) {
-        return AITools.playerInRange(entity, range);
+        return LevelUtils.playerInRange(entity, range);
     }
 }
